notebooks/requirements.txt
jupyterlab==3.2.1
pandas==1.3.4
numpy==1.20.3
scikit-learn==1.1.2
tensorflow-cpu==2.11.0
jupyterlab==3.2.1
matplotlib==3.4.3
